third_party/WebKit/LayoutTests/fast/dom/StyleSheet/css-insert-import-rule-to-shadow-stylesheets.html

<!doctype html>
<html>
<head>
<script>
if (window.testRunner)
    testRunner.waitUntilDone();

function runTest() {
    var sandbox = document.getElementById('sandbox');
    var host = document.createElement('p');
    var shadowRoot = host.createShadowRoot();
    shadowRoot.innerHTML = '<style>div { width: 100px; height: 100px; }</style><div></div>';
    sandbox.appendChild(host);
    document.body.offsetLeft;
    shadowRoot.styleSheets[0].insertRule('@import url(../resources/css-insert-import-rule.css);', 0);

    setTimeout(function(){
        if (window.testRunner)
            testRunner.notifyDone();
    }, 0);
}
</script>
</head>
<body onload="runTest()";>
<div id='sandbox'></div>
</body>
</html>
