third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/StyleSheet/css-insert-import-rule-to-shadow-stylesheets-expected.txt

CONSOLE ERROR: line 11: Uncaught TypeError: host.createShadowRoot is not a function
